Major refactoring for better organization an resuability.
